Implement BehaviorSubject for all associate call
The previous implementation was returning an observable to each caller.
That caused each caller to have their own copy of the data. Since
BehaviorSubjects are mutlicast this service will now hold the data and
notify components when the data is updated.

TODO: the call to fetch this data still comees from the component which
should not be the case, need to look into a better updating handle