Comments for the service redesign
This overhaul is going to change the way services work in the
application. Since, for times sake, we must remain backwards compatable
there are some strange behaviors and implementations.

This is where the comments are needed and will be added so that future
batches can see what has been done and what needs to be done.