Show spinner on loading
Since the BehaviorSubject has an inital value the isDataReady flag was
being set to true since the callback was being fired with the inital
value.

The fix was moving the setting of the flag into the loop for each
associate. This has redunant setting up the variable, however, once the
final change comes through to use AsyncSubjects this can be done in a
cleaner way.