Implement the BehaviorSubject for getByAssociateId
See 9644e6136d8465373bdcb0dd7164a604a77e0af3 for more detail. Note that
this change has a waterfall affect as it is the first updated method
that affects a reused componet.

The form componet needs to be throughly tested since the change there
will change the behavior for many pages and users. This method does not
have many side affects there but the interview call will when that
update comes through.