Update the updateAssociates method
This method is now using BehaviorSubjects. This has an error on the
associate page once an update is being sent due to a server 500 error.
The front end application is correctly responding to this error so this
should be ok. Once the backend is fixed it needs to be verified.