Updated Message for Comments
This will document for future batches what the shift in service design
is going toward so that if it is not finished they will have an idea of
why there are two ways the services are built.

This commit will mark v1 of the service redesign. Submitting PR for
through testing to ensure that this update did not produce unwanted side
effects.