Add the client list
This is using the fifty call to the client. Until the client service is
revamped the client list in the filter will only show fifty of them. The
clinet list will have to be a infinite scroll similar to the associate
pagnation that is being worked on by this branch.

The async subject changes could have introduced side effects in other
areas of the application need to investigate.

The direct usage of the observables.value is likely not a good practice
need to investigate.