Add show table in Associate page
The table is now showing as before in the associate page. This will
begin the update to allow for pagenation to optimize the load of the
page.

Made a styling change to the navbar so that the style for the page could
be determined. The nav bar was too large on the page to get a feel for
the desgin.